Make a command group
