+++ src/java/org/apache/cassandra/tools/nodetool/NodetoolCommand.java
@@ -207,7 +207,7 @@ import static org.apache.cassandra.tools.nodetool.Help.printTopCommandUsage;
                          TableStats.class,
                          TopPartitions.class,
                          TpStats.class,
-                         TrainCompressionDictionary.class,
+                         CompressionDictionary.class,
                          TruncateHints.class,
                          UpdateCIDRGroup.class,
                          UpgradeSSTable.class,
